+++ example/README.md
@@ -1,5 +1,5 @@
 This example was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
 
-It is linked to the react-form-builder package in the parent directory for development purposes.
+It is linked to the @onebeyond/react-form-builder package in the parent directory for development purposes.
 
 You can run `npm install` and then `npm start` to test your package.
