+++ rollup.config.js
@@ -6,30 +6,35 @@ import json from '@rollup/plugin-json'
 import replace from 'rollup-plugin-replace'
 import postcss from 'rollup-plugin-postcss'
 
-export default {
-  input: 'src/index.js',
-  output: {
-    name: 'ReactFormBuilder',
-    file: 'dist/index.js',
-    format: 'umd',
-    sourcemap: true,
+import pkg from './package.json'
 
-    globals: {
-      'react': 'React',
-      'theme-ui': 'theme-ui',
-      'react-markdown': 'react-markdown',
-      'react-phone-number-input': 'react-phone-number-input',
-      'react-hook-form': 'react-hook-form',
-      'react-datepicker': 'react-datepicker',
-      'react-select': 'react-select',
-      'react-scripts': 'react-scripts',
-      'react-dom': 'ReactDOM',
-      'react-ymd-date-select': 'react-ymd-date-select',
-      'countries-list': 'countries-list',
-      'i18n-iso-countries': 'i18n-iso-countries',
-    }
-  },
+/**
+ * Used for generating external dependencies
+ * Credit: Mateusz BurzyĹ„ski (https://github.com/Andarist)
+ * Source: https://github.com/rollup/rollup-plugin-babel/issues/148#issuecomment-399696316
+ */
+const makeExternalPredicate = (externalArr) => {
+  if (externalArr.length === 0) {
+    return () => false;
+  }
+  const pattern = new RegExp(`^(${externalArr.join('|')})($|/)`);
+  return (id) => pattern.test(id);
+};
 
+export default commandLineArgs => ({
+  input: pkg.source,
+  output: [
+    {
+      file: pkg.main,
+      format: 'cjs',
+      sourcemap: true,
+    },
+    {
+      file: pkg.module,
+      format: 'esm',
+      sourcemap: true,
+    },
+  ],
   plugins: [
     nodeResolve(),
     peerDepsExternal(),
@@ -39,21 +44,14 @@ export default {
     postcss({
       extensions: ['.css']
     }),
-    replace({ 'process.env.NODE_ENV': JSON.stringify('production') }),
+    ...(commandLineArgs.configDebug === true
+      ? []
+      : [replace({ 'process.env.NODE_ENV': JSON.stringify('production') })]),
   ],
 
-  external: [
-    'react',
-    'theme-ui',
-    'react-markdown',
-    'react-hook-form',
-    'react-phone-number-input',
-    'react-datepicker',
-    'react-scripts',
-    'react-select',
-    'react-dom',
-    'react-ymd-date-select',
-    'countries-list',
-    'i18n-iso-countries'
-  ]
-}
+  external: makeExternalPredicate([
+    // Handles both dependencies and peer dependencies so we don't have to manually maintain a list
+    ...Object.keys(pkg.dependencies || {}),
+    ...Object.keys(pkg.peerDependencies || {}),
+  ]),
+});
